backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Celery
redis==5.1.0
celery==5.4.0
slowapi==0.1.9

# AI & LangChain
openai==1.51.2
langchain==0.3.1
langchain-openai==0.2.1

# Authentication
pyjwt==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
cryptography==43.0.1

# HTTP Client
httpx==0.27.2
aiohttp==3.10.8
requests==2.32.3

# Image Storage
cloudinary==1.44.1

# Stripe
stripe==11.0.0

# Email
resend==2.4.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
faker==28.0.0

# Background Jobs
apscheduler==3.10.4

# OAuth & Authentication
authlib==1.3.1

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk==2.14.0

# HTTP Client (already present above but ensuring version)
httpx==0.27.2

# Development
black==24.10.0
flake8==7.1.1
mypy==1.11.2
